divide original and new proof test
